requirements.txt
wechat-clawbot
openai
python-dotenv
python-docx
pypdf
